tests: skip the double log failure test on 2.6
On 2.6 kernels this deadlocks.

+++ test/shell/lvconvert-repair-replace.sh
@@ -65,19 +65,21 @@ vgremove -ff $vg
 # FIXME  - exclusive activation for mirrors should work here
 test -e LOCAL_CLVMD && exit 0
 
-# 2-way, mirrored log
-# Double log failure, full replace
-vgcreate $vg "$dev1" "$dev2" "$dev3" "$dev4" "$dev5" "$dev6"
-lvcreate -aey --mirrorlog mirrored -m 1 --ig -L 1 -n 2way $vg \
-    "$dev1" "$dev2" "$dev3":0 "$dev4":0
-aux disable_dev "$dev3" "$dev4"
-lvconvert -y --repair $vg/2way 2>&1 | tee 2way.out
-lvs -a -o +devices $vg | not grep unknown
-not grep "WARNING: Failed" 2way.out
-vgreduce --removemissing $vg
-check mirror $vg 2way
-aux enable_dev "$dev3" "$dev4"
-vgremove -ff $vg
+if kernel_at_least 3 0 0; then
+	# 2-way, mirrored log
+	# Double log failure, full replace
+	vgcreate $vg "$dev1" "$dev2" "$dev3" "$dev4" "$dev5" "$dev6"
+	lvcreate -aey --mirrorlog mirrored -m 1 --ig -L 1 -n 2way $vg \
+	    "$dev1" "$dev2" "$dev3":0 "$dev4":0
+	aux disable_dev "$dev3" "$dev4"
+	lvconvert -y --repair $vg/2way 2>&1 | tee 2way.out
+	lvs -a -o +devices $vg | not grep unknown
+	not grep "WARNING: Failed" 2way.out
+	vgreduce --removemissing $vg
+	check mirror $vg 2way
+	aux enable_dev "$dev3" "$dev4"
+	vgremove -ff $vg
+fi
 
 # 3-way, mirrored log
 # Single log failure, replace
